Port `purchase_discount` to Odoo 14.0
Pretty much just made the tests pass.
