+++ purchase_discount/__manifest__.py
@@ -1,6 +1,7 @@
 # © 2004-2009 Tiny SPRL (<http://tiny.be>).
 # © 2014-2017 Tecnativa - Pedro M. Baeza
 # © 2016 ACSONE SA/NV (<http://acsone.eu>)
+# © 2020 Versada UAB
 # License AGPL-3 - See http://www.gnu.org/licenses/agpl-3.0.html
 {
     "name": "Purchase order lines with discounts",
@@ -9,8 +10,9 @@
     "Tecnativa, "
     "ACSONE SA/NV,"
     "GRAP,"
+    "Versada UAB,"
     "Odoo Community Association (OCA)",
-    "version": "13.0.1.0.1",
+    "version": "14.0.1.0.0",
     "category": "Purchase Management",
     "website": "https://github.com/OCA/purchase-workflow",
     "depends": ["purchase_stock"],
