+++ purchase_discount/models/__init__.py
@@ -1,5 +1,4 @@
 from . import purchase_order
 from . import stock_move
-from . import stock_rule
 from . import product_supplierinfo
 from . import res_partner
