+++ purchase_discount/views/res_partner_view.xml
@@ -17,7 +17,7 @@
                             name="open_commercial_entity"
                             type="object"
                             string="the parent company"
-                            class="oe_link"
+                            class="btn-link"
                         /></p>
                 </div>
             </xpath>
